perf: Improve speed of generating combinations
The function of generating combinations would not generate all the
possible cases.
Instead, it would generate the cases which are lower than the budget.
This saves a lot of time.
The progress bar of generating combinations is modified as well.
The maximum would not be based on the number of possible combinations,
hence the function of counting the combination number is removed.